change Firebase Data as singleton
